use `ref.midPoint` for color and opacity + move FILL/STROKE_CONFIG to src/mark
